requirements.txt
python-telegram-bot~=13.1
python-dotenv~=0.15.0